Create services to replace resolver dependency for Test and TestResult retrieval
