+++ server/resolvers/TestPlanVersion/testsResolver.js
@@ -1,3 +1,4 @@
+const getTests = require('../../models/services/TestsService');
 const ats = require('../../resources/ats.json');
 const commands = require('../../resources/commands.json');
 
@@ -9,27 +10,6 @@ const commands = require('../../resources/commands.json');
  * default atId for child fields as in the `renderableContent(atId: ID)` field.
  * @returns {array[*]} - An array of resolved tests.
  */
-const testsResolver = parentRecord => {
-    const isTestPlanVersion = !!parentRecord.tests;
-    const testPlanReport = isTestPlanVersion ? null : parentRecord;
-    const testPlanVersion = isTestPlanVersion
-        ? parentRecord
-        : testPlanReport.testPlanVersion;
-    const inferredAtId = testPlanReport?.atId;
-
-    // Populate nested At and Command fields
-    return testPlanVersion.tests.map(test => ({
-        ...test,
-        inferredAtId, // Not available in GraphQL, but used by child resolvers
-        ats: test.atIds.map(atId => ats.find(at => at.id === atId)),
-        scenarios: test.scenarios.map(scenario => ({
-            ...scenario,
-            at: ats.find(at => at.id === scenario.atId),
-            commands: scenario.commandIds.map(commandId => {
-                return commands.find(command => command.id === commandId);
-            })
-        }))
-    }));
-};
+const testsResolver = parentRecord => getTests(parentRecord);
 
 module.exports = testsResolver;
